Updates for 1.11 versions of the GD SDKs and Plugin
Updated for the 1.11 versions of the GD SDKs for Android and for iOS,
and of the GD PhoneGap Plugin.
Updated all the compatibility sections of the readme files.
Updated source for iOS to the latest practice.

+++ foriOS/src/appkinetics/core/gdDispatcher.m
@@ -1,4 +1,4 @@
-/* Copyright (c) 2014 Good Technology Corporation
+/* Copyright (c) 2015 Good Technology Corporation
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,7 +23,7 @@
 #import "gdRequest.h"
 
 typedef struct PROVIDER_LOOKUP {
-    int errorCode;
+    long errorCode;
     int index;
 } provider_lookup;
