+++ foriOS/src/pathstore/PathStore.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2014 Good Technology Corporation
+/* Copyright (c) 2015 Good Technology Corporation
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -30,7 +30,7 @@
 -(BOOL)isArray;
 -(BOOL)sameType:(NSObject *)that;
 -(instancetype)put:(NSObject *)o value:(NSObject *)value;
--(int)length;
+-(NSUInteger)length;
 -(id)pathGet:(id)path0, ...;
 -(id)pathGetv:(id)path0 arguments:(va_list)paths;
 -(id)pathGeta:(NSArray *)paths;
